requirements.txt
et-xmlfile==1.1.0
Jinja2==3.1.2
MarkupSafe==2.1.1
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
python-dateutil==2.8.2
pytz==2022.1
six==1.16.0
